requirements.txt
alembic==1.4.3
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
fastapi==0.61.1
FastAPI-SQLAlchemy==0.1.6
h11==0.9.0
httptools==0.1.1
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.6
pycparser==2.20
pydantic==1.6.1
PyJWT==2.0.1
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.6
Unipath==1.1
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
